Replaced the usage of ReadConsoleInputW and fgetwc with standard C++ input functions to make getchar32() work consistently in all environments, including cases when stdin is redirected.
The previous implementation of getchar32() relied on low-level console functions, which caused issues when running the code in subprocesses with redirected stdin. The ReadConsoleInputW function is designed to read from the console input buffer and may not function correctly with input redirection. Additionally, using fgetwc in a Windows environment could lead to potential issues in certain scenarios.

I encountered unexpected results when redirecting stdin, even without passing any argument. To replicate the issue, try the following command in your C# code:

```
            ProcessStartInfo startInfo = new()
            {
                FileName = ".\\main.exe",
                Arguments = "-m .\\7b-q4.bin -n 256 --repeat_penalty 1.0 --interactive-first -r \"User:\" -f .\\chat-with-bob.txt",
                RedirectStandardOutput = true,
                RedirectStandardInput = true,
                UseShellExecute = false,
                CreateNoWindow = true
            };
```

To address this problem and enable people to stream the program directly to a UI without worrying about the C++ part, I made the following changes:
- Replaced ReadConsoleInputW with std::getwchar(), a standard C++ input function that reads wide characters from std::wcin. This change allows getchar32() to handle both console and redirected stdin scenarios consistently.

With these modifications, getchar32() now functions as intended in various environments and ensures that console interactions work correctly, even when stdin is redirected.