Cleaned README installing orders for available options
The README has so many options for different OS and so on, and also multiple options for BLAS build, it takes time for the first time users to wrap head around.

As a simple fix I have updated the orders, of list. Instead of **Unordered list** I have provided the **Ordered list**.