Suggest setting AMDGPU_TARGETS
Currently, when building llama.cpp with CMake and hipBLAS option, it
will build for gfx900, gfx906, gfx908, gfx90a, gfx1030 if using 5.6.1 or
older. And, build for gfx906 if using ROCm 5.7.0 or 5.7.1.

The former raised multiple issues that CMake build wasn't working for
RDNA3 cards, and once ROCm 5.7.0 or 5.7.1 is available in popular linux
distributions, it will surely raise issues for most AMD cards owner.

This issue was fixed upstream 3 days ago, it will do something similar
to this suggestion.