cmake: fix clang build when CUDA is enabled (#4208)
Don't use the cxx_flags for .cu files so that -Wunreachable-code-break
and -Wunreachable-code-return would not be sent to to GCC which
doesn't understand them.
By default, nvcc uses gcc as the preprocessor even when Clang is used
for .c/.cpp.

This should fix the CUDA build with clang 14 (when gcc is used with nvcc).
With more complex configuration, it probably would be possible to use clang
itself (instead of nvcc) to compile CUDA code or use clang as the preprocessor for
nvcc. But these are probably to be quite rare use cases.
It seems that Clang 16 would require even more changes.