finetune: sort includes according to iwyu
This commit sorts the includes in finetune.cpp according to the
include-what-you-use tool.

The motivation for sorting is that it might make it easier in the
furture to just copy and paste the output from iwyu into the source
and also to make it easier to see what is missing.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>