nix: remove nixConfig from flake.nix
Why? I've demoed `nix run ggerganov/llama.cpp` to three people now, and
the prompts caused by this setting have been a fly in the ointment every
time, where I have to hand-wave away what's being asked... and because
it happens on every single command run, it's a major pain in the ass.

It's fine to leave it in the flake for those who read it, but the Nix
ecosystem isn't yet ready to have flakes with nixConfig, in my
judgement.